# public/openapi.yaml
openapi: 3.1.0
info:
  title: mindlog · id — API
  version: 0.2.2
  description: |
    Carte d'identité en ligne (1 page + API + MCP).

    ## Authentification
    Deux mécanismes, résolus par `currentIdentity()` :
    1. **Cookie de session** `mindlog_session` (HttpOnly, Secure, SameSite=Lax) —
       posé après login passkey ou échange clé→session. Préféré pour le navigateur.
       Pour les méthodes non sûres (POST/PUT/PATCH/DELETE), l'en-tête `Origin` est
       vérifié quand il est présent (protection CSRF).
    2. **Clé d'accès** — secret de 24 octets (base64url). Transmis soit par l'en-tête
       `X-Access-Key`, soit par le paramètre de requête `?key=...`. Utilisé par les
       liens privés `/k/{clé}`, le serveur MCP et le flux SSE.

    Les routes « contacts » exigent une relation **réciproque** entre les deux handles.
    Les routes « publiques » n'exigent rien (un viewer authentifié élargit la réponse).

    ## Sécurité transverse
    - **Rate-limiting** (par IP) : `429 Too Many Requests` + `Retry-After`.
      Backstop global de 300 req/min sur `/api/*` ; limites strictes sur
      `POST /api/identities` (10/h), `POST /api/recover` (5/15 min),
      `POST /api/passkeys/auth/*` (20/15 min), `GET /api/search` (60/min).
    - **Taille du corps** : `413` au-delà de 256 Ko (JSON) ou 55 Mo (multipart).
    - **CSRF** : pour une mutation authentifiée par cookie, l'`Origin` (ou à
      défaut le `Referer`) doit correspondre à l'origine de l'app.
...
servers:
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mindlog_session
    accessKeyHeader:
      type: apiKey
      in: header
      name: X-Access-Key
    accessKeyQuery:
      type: apiKey
      in: query
      name: key
    oauth2:
      type: oauth2
      description: OAuth 2.1 (Authorization Code + PKCE S256, DCR). mindlog est AS + RS.
      flows:
        authorizationCode:
          authorizationUrl: https://id.mindlog.today/oauth/authorize
          tokenUrl: https://id.mindlog.today/oauth/token
          refreshUrl: https://id.mindlog.today/oauth/token
          scopes:
            "mindlog:identity": Gérer la carte d'identité du compte connecté
  schemas:
...
    CardEvent:
      type: object
      properties:
        id: { type: integer }
        identity_id: { type: integer }
        title: { type: string }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time, nullable: true }
        location: { type: string }
        notes: { type: string }
        is_public: { type: integer, enum: [0, 1] }
...
security:
  - sessionCookie: []
  - accessKeyHeader: []
  - accessKeyQuery: []
paths:
...
  /api/session:
    get:
      tags: [Auth & Sessions]
      summary: État d'authentification léger (pour reconnexion auto)
      security:
        - sessionCookie: []
        - accessKeyQuery: []
      responses:
        '200':
          description: État
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated: { type: boolean }
                  handle: { type: string }
                  accessKey: { type: string }
                  displayName: { type: string, nullable: true }
                  hasPhoto: { type: boolean }
  /api/auth/session-from-key:
    post:
      tags: [Auth & Sessions]
      summary: Échanger une clé d'accès valide contre un cookie de session
      responses:
        '200': { description: Session créée }
        '401': { description: Clé invalide }
  /api/auth/pin:
    post:
      tags: [Auth & Sessions]
      summary: Générer un code PIN d'appairage (usage unique, TTL 10 min) pour connecter un nouvel appareil
      responses:
        '200': { description: 'PIN généré : { pin, expiresAt }' }
        '401': { description: Non authentifié }
  /api/auth/redeem-pin:
    post:
      tags: [Auth & Sessions]
      summary: Échanger un code PIN valide contre la clé d'accès (appareil non authentifié)
      security: []
      responses:
        '200': { description: 'PIN consommé : { accessKey, handle }' }
        '404': { description: PIN invalide ou expiré }
        '429': { description: Trop de tentatives }
...
  /api/auth/logout-all:
    post:
      tags: [Auth & Sessions]
      summary: Déconnecter tous les appareils
      responses:
        '200': { description: Toutes les sessions révoquées }
        '401': { description: Non authentifié }
  /api/sessions:
    get:
      tags: [Auth & Sessions]
      summary: Lister les sessions actives
      responses:
        '200':
          description: Sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, description: hash du token }
                        createdAt: { type: string }
                        lastSeen: { type: string }
                        userAgent: { type: string, nullable: true }
                        current: { type: boolean }
        '401': { description: Non authentifié }
  /api/sessions/{id}:
    delete:
      tags: [Auth & Sessions]
      summary: Révoquer une session précise (par son hash)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Révoquée }
        '401': { description: Non authentifié }
  /api/access-key/rotate:
    post:
      tags: [Auth & Sessions]
      summary: Régénérer la clé d'accès (invalide l'ancienne + toutes les sessions)
      responses:
        '200':
          description: Nouvelle clé
          content:
            application/json:
              schema: { type: object, properties: { accessKey: { type: string } } }
        '401': { description: Non authentifié }
...
  /api/passkeys:
    get:
      tags: [Passkeys]
      summary: Lister les passkeys enregistrées
      responses:
        '200': { description: Liste }
        '401': { description: Non authentifié }
  /api/passkeys/register/begin:
    post:
      tags: [Passkeys]
      summary: Démarrer l'enregistrement d'une passkey (options WebAuthn)
      responses:
        '200': { description: PublicKeyCredentialCreationOptions }
        '401': { description: Non authentifié }
  /api/passkeys/register/finish:
    post:
      tags: [Passkeys]
      summary: Finaliser l'enregistrement d'une passkey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                name: { type: string }
                response: { type: object, description: RegistrationResponseJSON }
      responses:
        '200': { description: Passkey enregistrée }
        '400': { description: Échec de vérification }
        '401': { description: Non authentifié }
  /api/passkeys/{credId}:
    delete:
      tags: [Passkeys]
      summary: Supprimer une passkey
      parameters:
        - { name: credId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Supprimée }
        '401': { description: Non authentifié }
  /api/passkeys/auth/begin:
    post:
      tags: [Passkeys]
      summary: Démarrer l'authentification par passkey
      security: []
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [handle], properties: { handle: { type: string } } }
      responses:
        '200': { description: PublicKeyCredentialRequestOptions }
        '400': { description: Handle requis ou inconnu }
  /api/passkeys/auth/finish:
    post:
      tags: [Passkeys]
      summary: Finaliser l'authentification par passkey (ouvre une session)
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [handle, response]
              properties:
                handle: { type: string }
                response: { type: object, description: AuthenticationResponseJSON }
      responses:
        '200': { description: Connecté (cookie posé) }
        '401': { description: Échec d'authentification }
...
  /api/agenda:
    post:
      tags: [Agenda & Disponibilité]
      summary: Ajouter un événement à l'agenda
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title, starts_at]
              properties:
                title: { type: string }
                starts_at: { type: string, format: date-time }
                ends_at: { type: string, nullable: true }
                location: { type: string }
                notes: { type: string }
                is_public: { type: boolean }
      responses:
        '201': { description: Événement créé, content: { application/json: { schema: { $ref: '#/components/schemas/CardEvent' } } } }
        '400': { description: title et starts_at requis }
        '401': { description: Non authentifié }
  /api/agenda/{id}:
    delete:
      tags: [Agenda & Disponibilité]
      summary: Supprimer un événement
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Supprimé }
        '401': { description: Non authentifié }
        '404': { description: Introuvable }
  /api/availability/{day}:
    put:
      tags: [Agenda & Disponibilité]
      summary: Définir la disponibilité d'un jour (free/busy)
      parameters:
        - { name: day, in: path, required: true, schema: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' } }
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { status: { type: string, enum: [free, busy] } } }
      responses:
        '200': { description: Enregistré }
        '400': { description: Date ou statut invalide }
        '401': { description: Non authentifié }
...